pyproject.toml
[project]
name = "ticketless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    { name = "kwasa", email = "kwasa@ticketless.io" },
     { name = "david", email = "david@ticketless.io" }
]
maintainers = [
    { name = "kwasa", email = "kwasa@ticketless.io" },
     { name = "david", email = "david@ticketless.io" }
]
keywords = [
    "django",
    "ticket",
    "management",
    "system",
    "multi-tenant",
    "ticketless"
]
dependencies = [
    "boto3>=1.41.5",
    "celery>=5.5.3",
    "coverage>=7.12.0",
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.8",
    "django-allauth[socialaccount]>=65.13.1",
    "django-browser-reload>=1.21.0",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.9.0",
    "django-crispy-forms>=2.5",
    "django-debug-toolbar>=6.1.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-guardian>=3.2.0",
    "django-qr-code>=4.2.0",
    "django-redis>=6.0.0",
    "django-storages>=1.14.6",
    "django-tenants>=3.9.0",
    "django-weasyprint>=2.4.0",
    "django-webpack-loader>=3.2.2",
    "django-widget-tweaks>=1.5.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djlint>=1.36.4",
    "djoser>=2.3.3",
    "drf-spectacular>=0.29.0",
    "flower>=2.0.1",
    "hiredis>=3.3.0",
    "mypy>=1.19.0",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "pytest>=9.0.1",
    "pytest-django>=4.11.1",
    "qrcode[pil]>=8.2",
    "rav>=0.2.0",
    "redis>=7.1.0",
    "ruff>=0.14.7",
    "sentry-sdk>=2.46.0",
    "sphinx>=8.2.3",
    "sphinx-autobuild>=2025.8.25",
    "whitenoise>=6.11.0",
]



# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=config.settings.test --reuse-db --import-mode=importlib"
python_files = [
    "tests.py",
    "test_*.py",
]

# ==== Coverage ====
[tool.coverage.run]
include = ["ticketless/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.12"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/djlint/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
target-version = "py312"
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    "*/migrations/*.py",
    "staticfiles/*",
]

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TC",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
"S101", # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
"RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
"SIM102", # sometimes it's better to nest
"COM812"
]
# The fixes in extend-unsafe-fixes will require
# provide the `--unsafe-fixes` flag when fixing.
extend-unsafe-fixes = [
  "SIM101", # simplify nested if statements
  "SIM201", # simplify multiple comparisons
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "yt-dlp>=2025.11.12",
]
